pytorch: renamed from libtorch
